Slider now supports different animations
Rather than pass an `animated` flag, the slider now has an enum for setting the type of animation. This property is passed to the popUp view when a request is made to show/hide.

The bounce animation was also updated to use the newer spring damping API. Since this works on the UIView level rather than the underlying CALayer, `actionForLayer:forKey:` was removed (since that will nullify the animation). Removing that method didn't seem to have any affect on the animations in the example app.